+++ src/main/java/com/woosan/hr_system/auth/service/CustomAuthenticationSuccessHandler.java
@@ -1,6 +1,6 @@
 package com.woosan.hr_system.auth.service;
 
-import com.woosan.hr_system.auth.dao.PasswordDAO;
+import com.woosan.hr_system.auth.dao.PasswordDAOImpl;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.extern.slf4j.Slf4j;
@@ -16,7 +16,7 @@ import java.io.IOException;
 public class CustomAuthenticationSuccessHandler implements AuthenticationSuccessHandler {
 
     @Autowired
-    private PasswordDAO passwordDAO;
+    private PasswordDAOImpl passwordDAO;
 
     @Override
     public void onAuthenticationSuccess(HttpServletRequest request, HttpServletResponse response, Authentication authentication) throws IOException {
