+++ src/main/resources/mybatis/mappers/PasswordMapper.xml
@@ -2,7 +2,7 @@
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
-<mapper namespace="com.woosan.hr_system.auth.dao.PasswordDAO">
+<mapper namespace="password">
     <resultMap id="passwordResultMap" type="com.woosan.hr_system.auth.model.Password">
         <id property="employeeId" column="employee_id"/>
         <result property="password" column="password"/>
